backend/requirements_test.txt
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
coverage==7.4.1
freezegun==1.4.0
model-bakery==1.17.0
django-filter==23.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-storages==1.14.2
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
paypalrestsdk==1.13.1
djangorestframework==3.14.0 